add GPA calculating service in separate folder

+++ GPACalculator.API/Db/Entities/SubjectEntity.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; } 
         public string Name { get; set; }
 
-        public double Credit { get; set; }
+        public int Credit { get; set; }
 
  
     }
