+++ GPACalculator.API/Models/Requests/CreateSubjectRequest.cs
@@ -3,6 +3,6 @@
     public class CreateSubjectRequest
     {
         public  string Name { get; set; }
-        public double Credit { get; set; }
+        public int Credit { get; set; }
     }
 }
